working fine

+++ app/src/main/java/com/ihsan/cricplanet/model/fixture/TeamDlData.kt
@@ -5,9 +5,9 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class TeamDlData(
-    val overs: Float?,
+    val overs: Double?,
     val score: Int?,
-    val wickets_out: Int?
+    val wickets_out: Double?
 ):Parcelable
 {
     constructor():this(
